misc/HtmlTestSuite/results/test-0866.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <img src="./unoptimized-image.jpg"/>
    <script>
var check_report_format = (reports, observer) =&gt; {
  let report = reports[0];
  assert_equals(report.type, "document-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "lossy-images-max-bpp");
  assert_equals(report.body.disposition, "enforce");
  assert_equals(report.body.sourceFile, document.getElementsByTagName('img')[0].src);
  assert_equals(report.body.lineNumber, null);
  assert_equals(report.body.columnNumber, null);
};

async_test(t =&gt; {
  new ReportingObserver(t.step_func_done(check_report_format),
                        {types: ['document-policy-violation'], buffered: true}).observe();
}, "unoptimized-images Report Format");
    </script>
  </body>
</html>
